+++ resources/views/layouts/app.blade.php
@@ -137,6 +137,7 @@
     <script src="{{ \App\Support\Assets::url('assets/js/lightbox.js') }}" defer></script>
     <script src="{{ \App\Support\Assets::url('assets/js/infinite-scroll.js') }}" defer></script>
     <script src="{{ \App\Support\Assets::url('assets/js/post-menu.js') }}" defer></script>
+    <script src="{{ \App\Support\Assets::url('assets/js/like.js') }}" defer></script>
     <script src="{{ \App\Support\Assets::url('assets/js/header-panels.js') }}" defer></script>
     <script src="{{ \App\Support\Assets::url('assets/js/notifications-live.js') }}" defer></script>
     <script src="{{ \App\Support\Assets::url('assets/js/compose-shortcut.js') }}" defer></script>